JAVA-8360 Split or move spring-5-security module
