+++ spring-security-modules/spring-5-security/README.md
@@ -4,7 +4,6 @@ This module contains articles about Spring Security 5
 
 ## Relevant articles:
 
-- [Extra Login Fields with Spring Security](https://www.baeldung.com/spring-security-extra-login-fields)
 - [A Custom Spring SecurityConfigurer](https://www.baeldung.com/spring-security-custom-configurer)
 - [New Password Storage In Spring Security 5](https://www.baeldung.com/spring-security-5-password-storage)
 - [Default Password Encoder in Spring Security 5](https://www.baeldung.com/spring-security-5-default-password-encoder)
