--- spring-security-modules/spring-5-security/src/main/java/com/baeldung/loginextrafieldscustom/CustomAuthenticationToken.java
@@ -1,28 +1,0 @@
-package com.baeldung.loginextrafieldscustom;
-
-import java.util.Collection;
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-public class CustomAuthenticationToken extends UsernamePasswordAuthenticationToken {
-
-    private String domain;
-
-    public CustomAuthenticationToken(Object principal, Object credentials, String domain) {
-        super(principal, credentials);
-        this.domain = domain;
-        super.setAuthenticated(false);
-    }
-
-    public CustomAuthenticationToken(Object principal, Object credentials, String domain, 
-        Collection<? extends GrantedAuthority> authorities) {
-        super(principal, credentials, authorities);
-        this.domain = domain;
-        super.setAuthenticated(true); // must use super, as we override
-    }
-
-    public String getDomain() {
-        return this.domain;
-    }
-}
