--- spring-security-modules/spring-5-security/src/main/java/com/baeldung/loginextrafieldscustom/CustomUserDetailsService.java
@@ -1,10 +1,0 @@
-package com.baeldung.loginextrafieldscustom;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-public interface CustomUserDetailsService {
-
-    UserDetails loadUserByUsernameAndDomain(String username, String domain) throws UsernameNotFoundException;
-
-}
