--- spring-security-modules/spring-5-security/src/main/java/com/baeldung/loginextrafieldscustom/UserRepository.java
@@ -1,7 +1,0 @@
-package com.baeldung.loginextrafieldscustom;
-
-public interface UserRepository {
-
-    public User findUser(String username, String domain);
-    
-}
