--- spring-security-modules/spring-5-security/src/main/java/com/baeldung/loginextrafieldssimple/SimpleAuthenticationFilter.java
@@ -1,54 +1,0 @@
-package com.baeldung.loginextrafieldssimple;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.security.authentication.AuthenticationServiceException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-public class SimpleAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
-
-    public static final String SPRING_SECURITY_FORM_DOMAIN_KEY = "domain";
-
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) 
-        throws AuthenticationException {
-
-        if (!request.getMethod()
-            .equals("POST")) {
-            throw new AuthenticationServiceException("Authentication method not supported: " + request.getMethod());
-        }
-
-        UsernamePasswordAuthenticationToken authRequest = getAuthRequest(request);
-        setDetails(request, authRequest);
-        return this.getAuthenticationManager()
-            .authenticate(authRequest);
-    }
-
-    private UsernamePasswordAuthenticationToken getAuthRequest(HttpServletRequest request) {
-        String username = obtainUsername(request);
-        String password = obtainPassword(request);
-        String domain = obtainDomain(request);
-
-        if (username == null) {
-            username = "";
-        }
-        if (password == null) {
-            password = "";
-        }
-        if (domain == null) {
-            domain = "";
-        }
-
-        String usernameDomain = String.format("%s%s%s", username.trim(), 
-            String.valueOf(Character.LINE_SEPARATOR), domain);
-        return new UsernamePasswordAuthenticationToken(usernameDomain, password);        
-    }
-
-    private String obtainDomain(HttpServletRequest request) {
-        return request.getParameter(SPRING_SECURITY_FORM_DOMAIN_KEY);
-    }
-}
