+++ spring-security-modules/spring-security-web-login/src/main/java/com/baeldung/loginextrafieldscustom/UserRepository.java
@@ -1,0 +1,7 @@
+package com.baeldung.loginextrafieldscustom;
+
+public interface UserRepository {
+
+    public User findUser(String username, String domain);
+    
+}
