+++ spring-security-modules/spring-security-web-login/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+server.port=8081
+
+logging.level.root=INFO
+
+logging.level.com.baeldung.dsl.ClientErrorLoggingFilter=DEBUG
+
+logging.level.org.springframework.security=DEBUG
