+++ spring-security-modules/spring-security-web-login/src/main/resources/static/css/main.css
